根据数据库修改了一些数据的数据类型

+++ src/main/java/com/guotu/gt/controller/PermissionMenu1Controller.java
@@ -33,14 +33,14 @@ public class PermissionMenu1Controller {
 
     @DeleteMapping
     @ApiOperation(value = "根据code删除一个一级菜单信息")
-    public Result<Object> delete(@RequestParam("code") Byte code){
+    public Result<Object> delete(@RequestParam("code") int code){
         permissionMenu1Service.delete(code);
         return ResultUtil.success();
     }
 
     @GetMapping("/findByCode")
     @ApiOperation(value = "根据code查询一级菜单信息")
-    public Result<PermissionMenu1> findByCode(@RequestParam("code") Byte code){
+    public Result<PermissionMenu1> findByCode(@RequestParam("code") int code){
         return ResultUtil.success(permissionMenu1Service.findByCode(code));
     }
